One file per class for WeakPtrImpl, WeakPtrFactory, and CanMakeWeakPtr
https://bugs.webkit.org/show_bug.cgi?id=273084
rdar://126878970

Reviewed by Chris Dumez.

* Source/WTF/wtf/CanMakeWeakPtr.h:
* Source/WTF/wtf/WeakPtr.h:
* Source/WTF/wtf/WeakPtrImpl.h:

Canonical link: https://commits.webkit.org/277848@main
