+++ Source/WTF/wtf/WeakPtr.h
@@ -26,13 +26,13 @@
 
 #pragma once
 
+#include <wtf/CanMakeWeakPtr.h>
 #include <wtf/CompactRefPtrTuple.h>
 #include <wtf/Packed.h>
+#include <wtf/WeakPtrFactory.h>
 #include <wtf/WeakRef.h>
 
 namespace WTF {
-
-template<typename, typename> class WeakPtrFactory;
 
 template<typename, typename, typename = DefaultWeakPtrImpl> class WeakHashMap;
 template<typename, typename = DefaultWeakPtrImpl, EnableWeakPtrThreadingAssertions = EnableWeakPtrThreadingAssertions::Yes> class WeakHashSet;
@@ -167,212 +167,6 @@ private:
     bool m_shouldEnableAssertions { true };
 #endif
 };
-
-// Note: you probably want to inherit from CanMakeWeakPtr rather than use this directly.
-template<typename T, typename WeakPtrImpl = DefaultWeakPtrImpl>
-class WeakPtrFactory {
-    WTF_MAKE_NONCOPYABLE(WeakPtrFactory);
-    WTF_MAKE_FAST_ALLOCATED;
-public:
-    using ObjectType = T;
-    using WeakPtrImplType = WeakPtrImpl;
-
-    WeakPtrFactory()
-#if ASSERT_ENABLED
-        : m_wasConstructedOnMainThread(isMainThread())
-#endif
-    {
-    }
-
-    void prepareForUseOnlyOnNonMainThread()
-    {
-#if ASSERT_ENABLED
-        ASSERT(m_wasConstructedOnMainThread);
-        m_wasConstructedOnMainThread = false;
-#endif
-    }
-
-    ~WeakPtrFactory()
-    {
-        if (m_impl)
-            m_impl->clear();
-    }
-
-    WeakPtrImpl* impl() const
-    {
-        return m_impl.get();
-    }
-
-    void initializeIfNeeded(const T& object) const
-    {
-        if (m_impl)
-            return;
-
-        ASSERT(m_wasConstructedOnMainThread == isMainThread());
-
-        static_assert(std::is_final_v<WeakPtrImpl>);
-        m_impl = adoptRef(*new WeakPtrImpl(const_cast<T*>(&object)));
-    }
-
-    template<typename U> WeakPtr<U, WeakPtrImpl> createWeakPtr(U& object, EnableWeakPtrThreadingAssertions enableWeakPtrThreadingAssertions = EnableWeakPtrThreadingAssertions::Yes) const
-    {
-        initializeIfNeeded(object);
-
-        ASSERT(&object == m_impl->template get<T>());
-        return WeakPtr<U, WeakPtrImpl>(*m_impl, enableWeakPtrThreadingAssertions);
-    }
-
-    void revokeAll()
-    {
-        if (RefPtr impl = std::exchange(m_impl, nullptr))
-            impl->clear();
-    }
-
-    unsigned weakPtrCount() const
-    {
-        return m_impl ? m_impl->refCount() - 1 : 0u;
-    }
-
-#if ASSERT_ENABLED
-    bool isInitialized() const { return m_impl; }
-#endif
-
-private:
-    template<typename, typename, EnableWeakPtrThreadingAssertions> friend class WeakHashSet;
-    template<typename, typename, EnableWeakPtrThreadingAssertions> friend class WeakListHashSet;
-    template<typename, typename, typename> friend class WeakHashMap;
-    template<typename, typename, typename> friend class WeakPtr;
-    template<typename, typename> friend class WeakRef;
-
-    mutable RefPtr<WeakPtrImpl> m_impl;
-#if ASSERT_ENABLED
-    bool m_wasConstructedOnMainThread;
-#endif
-};
-
-// Note: you probably want to inherit from CanMakeWeakPtrWithBitField rather than use this directly.
-template<typename T, typename WeakPtrImpl = DefaultWeakPtrImpl>
-class WeakPtrFactoryWithBitField {
-    WTF_MAKE_NONCOPYABLE(WeakPtrFactoryWithBitField);
-    WTF_MAKE_FAST_ALLOCATED;
-public:
-    using ObjectType = T;
-    using WeakPtrImplType = WeakPtrImpl;
-
-    WeakPtrFactoryWithBitField()
-#if ASSERT_ENABLED
-        : m_wasConstructedOnMainThread(isMainThread())
-#endif
-    {
-    }
-
-    ~WeakPtrFactoryWithBitField()
-    {
-        if (auto* pointer = m_impl.pointer())
-            pointer->clear();
-    }
-
-    WeakPtrImpl* impl() const
-    {
-        return m_impl.pointer();
-    }
-
-    void initializeIfNeeded(const T& object) const
-    {
-        if (m_impl.pointer())
-            return;
-
-        ASSERT(m_wasConstructedOnMainThread == isMainThread());
-
-        static_assert(std::is_final_v<WeakPtrImpl>);
-        m_impl.setPointer(adoptRef(*new WeakPtrImpl(const_cast<T*>(&object))));
-    }
-
-    template<typename U> WeakPtr<U, WeakPtrImpl> createWeakPtr(U& object, EnableWeakPtrThreadingAssertions enableWeakPtrThreadingAssertions = EnableWeakPtrThreadingAssertions::Yes) const
-    {
-        initializeIfNeeded(object);
-
-        ASSERT(&object == m_impl.pointer()->template get<T>());
-        return WeakPtr<U, WeakPtrImpl>(*m_impl.pointer(), enableWeakPtrThreadingAssertions);
-    }
-
-    void revokeAll()
-    {
-        if (auto* pointer = m_impl.pointer()) {
-            pointer->clear();
-            m_impl.setPointer(nullptr);
-        }
-    }
-
-    unsigned weakPtrCount() const
-    {
-        if (auto* pointer = m_impl.pointer())
-            return pointer->refCount() - 1;
-        return 0;
-    }
-
-#if ASSERT_ENABLED
-    bool isInitialized() const { return m_impl.pointer(); }
-#endif
-
-    uint16_t bitfield() const { return m_impl.type(); }
-    void setBitfield(uint16_t value) const { return m_impl.setType(value); }
-
-private:
-    template<typename, typename, EnableWeakPtrThreadingAssertions> friend class WeakHashSet;
-    template<typename, typename, EnableWeakPtrThreadingAssertions> friend class WeakListHashSet;
-    template<typename, typename, typename> friend class WeakHashMap;
-    template<typename, typename, typename> friend class WeakPtr;
-    template<typename, typename> friend class WeakRef;
-
-    mutable CompactRefPtrTuple<WeakPtrImpl, uint16_t> m_impl;
-#if ASSERT_ENABLED
-    bool m_wasConstructedOnMainThread;
-#endif
-};
-
-// We use lazy initialization of the WeakPtrFactory by default to avoid unnecessary initialization. Eager
-// initialization is however useful if you plan to call construct WeakPtrs from other threads.
-enum class WeakPtrFactoryInitialization { Lazy, Eager };
-
-template<typename WeakPtrFactoryType, WeakPtrFactoryInitialization initializationMode = WeakPtrFactoryInitialization::Lazy>
-class CanMakeWeakPtrBase {
-public:
-    using WeakValueType = typename WeakPtrFactoryType::ObjectType;
-    using WeakPtrImplType = typename WeakPtrFactoryType::WeakPtrImplType;
-
-    const WeakPtrFactoryType& weakPtrFactory() const { return m_weakPtrFactory; }
-    WeakPtrFactoryType& weakPtrFactory() { return m_weakPtrFactory; }
-
-protected:
-    CanMakeWeakPtrBase()
-    {
-        if (initializationMode == WeakPtrFactoryInitialization::Eager)
-            initializeWeakPtrFactory();
-    }
-
-    CanMakeWeakPtrBase(const CanMakeWeakPtrBase&)
-    {
-        if (initializationMode == WeakPtrFactoryInitialization::Eager)
-            initializeWeakPtrFactory();
-    }
-
-    CanMakeWeakPtrBase& operator=(const CanMakeWeakPtrBase&) { return *this; }
-
-    void initializeWeakPtrFactory()
-    {
-        m_weakPtrFactory.initializeIfNeeded(static_cast<WeakValueType&>(*this));
-    }
-
-private:
-    WeakPtrFactoryType m_weakPtrFactory;
-};
-
-template<typename T, WeakPtrFactoryInitialization initializationMode = WeakPtrFactoryInitialization::Lazy, typename WeakPtrImpl = DefaultWeakPtrImpl>
-using CanMakeWeakPtr = CanMakeWeakPtrBase<WeakPtrFactory<T, WeakPtrImpl>, initializationMode>;
-
-template<typename T, WeakPtrFactoryInitialization initializationMode = WeakPtrFactoryInitialization::Lazy, typename WeakPtrImpl = DefaultWeakPtrImpl>
-using CanMakeWeakPtrWithBitField = CanMakeWeakPtrBase<WeakPtrFactoryWithBitField<T, WeakPtrImpl>, initializationMode>;
 
 template<typename T, typename U, typename WeakPtrImpl> inline WeakPtrImpl* weak_ptr_impl_cast(WeakPtrImpl* impl)
 {
@@ -528,9 +322,6 @@ WeakPtr(const RefPtr<T>& value, EnableWeakPtrThreadingAssertions = EnableWeakPtr
 template<typename T, typename PtrTraits = RawPtrTraits<SingleThreadWeakPtrImpl>> using SingleThreadWeakPtr = WeakPtr<T, SingleThreadWeakPtrImpl, PtrTraits>;
 template<typename T> using SingleThreadPackedWeakPtr = WeakPtr<T, SingleThreadWeakPtrImpl, PackedPtrTraits<SingleThreadWeakPtrImpl>>;
 
-template<typename T, WeakPtrFactoryInitialization initializationMode = WeakPtrFactoryInitialization::Lazy>
-using CanMakeSingleThreadWeakPtr = CanMakeWeakPtr<T, initializationMode, SingleThreadWeakPtrImpl>;
-
 template<typename T, EnableWeakPtrThreadingAssertions enableWeakPtrThreadingAssertions = EnableWeakPtrThreadingAssertions::Yes>
 using SingleThreadWeakHashSet = WeakHashSet<T, SingleThreadWeakPtrImpl, enableWeakPtrThreadingAssertions>;
 
@@ -541,9 +332,6 @@ using SingleThreadWeakListHashSet = WeakListHashSet<T, SingleThreadWeakPtrImpl, 
 
 } // namespace WTF
 
-using WTF::CanMakeWeakPtr;
-using WTF::CanMakeWeakPtrWithBitField;
-using WTF::CanMakeSingleThreadWeakPtr;
 using WTF::SingleThreadPackedWeakPtr;
 using WTF::SingleThreadWeakPtr;
 using WTF::SingleThreadWeakHashSet;
@@ -553,5 +341,3 @@ using WTF::SingleThreadWeakHashMap;
 using WTF::WeakHashSet;
 using WTF::WeakListHashSet;
 using WTF::WeakPtr;
-using WTF::WeakPtrFactory;
-using WTF::WeakPtrFactoryInitialization;
